+++ OnchainClob.Sdk/Trading/Requests/PlaceOrderRequest.cs
@@ -1,12 +1,13 @@
 ﻿using OnchainClob.Common;
+using System.Numerics;
 
 namespace OnchainClob.Trading.Requests
 {
     public class PlaceOrderRequest : ITraderRequest
     {
         public int Priority => 0;
-        public decimal Price { get; init; }
-        public decimal Qty { get; init; }
+        public BigInteger Price { get; init; }
+        public BigInteger Qty { get; init; }
         public Side Side { get; init; }
         public ulong OrderId => Side == Side.Buy ? 0u : 1u;
     }
